@@ -110,17 +110,17 @@
 	],
 	"dependencies": {
 		"@agm/core": "^1.0.0-beta.6",
-		"@angular/animations": "^8.2.4",
-		"@angular/cdk": "^8.1.4",
-		"@angular/common": "^8.2.4",
-		"@angular/compiler": "^8.2.4",
-		"@angular/core": "^8.2.4",
-		"@angular/forms": "^8.2.4",
-		"@angular/platform-browser": "^8.2.4",
-		"@angular/platform-browser-dynamic": "^8.2.4",
-		"@angular/pwa": "~0.803.2",
-		"@angular/router": "^8.2.4",
-		"@angular/service-worker": "^8.2.4",
+		"@akveo/ng2-completer": "^9.0.1",
+		"@angular/animations": "^9.1.7",
+		"@angular/cdk": "^9.1.7",
+		"@angular/common": "^9.1.7",
+		"@angular/compiler": "^9.1.7",
+		"@angular/core": "^9.1.7",
+		"@angular/forms": "^9.1.7",
+		"@angular/platform-browser": "^9.1.7",
+		"@angular/platform-browser-dynamic": "^9.1.7",
+		"@angular/router": "^9.1.7",
+		"@angular/service-worker": "^9.1.7",
 		"@asymmetrik/ngx-leaflet": "^5.0.2",
 		"@cloudinary/angular-5.x": "^1.0.3",
 		"@datorama/akita": "^4.23.0",
@@ -285,31 +284,29 @@
 		"karma-coverage-istanbul-reporter": "^2.0.5",
 		"karma-jasmine": "^2.0.1",
 		"karma-jasmine-html-reporter": "^1.4.2",
-		"lerna": "^3.15.0",
-		"lerna-changelog": "^0.8.2",
+		"lerna": "^3.21.0",
+		"lerna-changelog": "^1.0.1",
 		"lighthouse": "^5.0.0",
 		"lint-staged": "^9.2.1",
-		"ng-packagr": "^5.5.0",
+		"ng-packagr": "^9.1.5",
 		"npm-run-all": "^4.1.5",
-		"prettier": "^1.18.2",
+		"prettier": "^2.0.5",
 		"prettier-tslint": "^0.4.2",
-		"pretty-quick": "^1.10.0",
+		"pretty-quick": "^2.0.1",
 		"protractor": "^5.4.2",
-		"rimraf": "^3.0.0",
+		"rimraf": "^3.0.2",
 		"rxjs-tslint": "0.1.5",
-		"semantic-release": "^15.13.24",
-		"snyk": "^1.200.1",
-		"stylelint": "^10.1.0",
+		"semantic-release": "^17.0.7",
+		"snyk": "^1.323.2",
+		"stylelint": "^13.5.0",
 		"ts-jest": "^24.0.2",
-		"ts-loader": "^6.0.4",
-		"ts-node": "^8.3.0",
-		"tsconfig-paths": "^3.8.0",
-		"tsickle": "^0.37.0",
-		"tslint": "^5.19.0",
+		"ts-loader": "^7.0.4",
+		"ts-node": "^8.10.1",
+		"tsconfig-paths": "^3.9.0",
+		"tslint": "^6.1.2",
 		"tslint-config-prettier": "^1.18.0",
-		"tslint-language-service": "^0.9.9",
-		"typescript": "~3.5.3",
-		"webpack-bundle-analyzer": "^3.4.1"
+		"typescript": "~3.8.3",
+		"webpack-bundle-analyzer": "^3.8.0"
 	},
 	"engines": {
 		"node": ">=10.15.0",
