@@ -1,71 +1,66 @@
-import { ModuleWithProviders, NgModule, Optional, SkipSelf } from '@angular/core';
+import {
+	ModuleWithProviders,
+	NgModule,
+	Optional,
+	SkipSelf
+} from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NbAuthModule } from '@nebular/auth';
 import { NbSecurityModule, NbRoleProvider } from '@nebular/security';
 import { of as observableOf } from 'rxjs';
 
 import { throwIfAlreadyLoaded } from './module-import-guard';
-import {
-  AnalyticsService,
-  LayoutService
-} from './utils';
+import { AnalyticsService, LayoutService } from './utils';
 import { AuthModule } from './auth/auth.module';
 
-const DATA_SERVICES = [
-];
+const DATA_SERVICES = [];
 
 export class NbSimpleRoleProvider extends NbRoleProvider {
-  getRole() {
-    // here you could provide any role based on any auth flow
-    return observableOf('guest');
-  }
+	getRole() {
+		// here you could provide any role based on any auth flow
+		return observableOf('guest');
+	}
 }
 
 export const NB_CORE_PROVIDERS = [
-  ...DATA_SERVICES,
+	...DATA_SERVICES,
 
-  NbSecurityModule.forRoot({
-    accessControl: {
-      guest: {
-        view: '*',
-      },
-      user: {
-        parent: 'guest',
-        create: '*',
-        edit: '*',
-        remove: '*',
-      },
-    },
-  }).providers,
+	NbSecurityModule.forRoot({
+		accessControl: {
+			guest: {
+				view: '*'
+			},
+			user: {
+				parent: 'guest',
+				create: '*',
+				edit: '*',
+				remove: '*'
+			}
+		}
+	}).providers,
 
-  {
-    provide: NbRoleProvider, useClass: NbSimpleRoleProvider,
-  },
-  AnalyticsService,
-  LayoutService
+	{
+		provide: NbRoleProvider,
+		useClass: NbSimpleRoleProvider
+	},
+	AnalyticsService,
+	LayoutService
 ];
 
 @NgModule({
-  imports: [
-    CommonModule,
-    AuthModule
-  ],
-  exports: [
-    NbAuthModule
-  ],
-  declarations: [],
+	imports: [CommonModule, AuthModule],
+	exports: [NbAuthModule],
+	declarations: []
 })
 export class CoreModule {
-  constructor(@Optional() @SkipSelf() parentModule: CoreModule) {
-    throwIfAlreadyLoaded(parentModule, 'CoreModule');
-  }
+	constructor(@Optional() @SkipSelf() parentModule: CoreModule) {
+		throwIfAlreadyLoaded(parentModule, 'CoreModule');
+	}
 
-  static forRoot(): ModuleWithProviders {
-    return <ModuleWithProviders>{
-      ngModule: CoreModule,
-      providers: [
-        ...NB_CORE_PROVIDERS,
-      ],
-    };
-  }
+	static forRoot(): ModuleWithProviders<CoreModule> {
+		return {
+			ngModule: CoreModule,
+			providers: [...NB_CORE_PROVIDERS]
+		};
+	}
 }
