@@ -117,8 +117,8 @@ const PIPES = [
 	]
 })
 export class ThemeModule {
-	static forRoot(): ModuleWithProviders {
-		return <ModuleWithProviders>{
+	static forRoot(): ModuleWithProviders<ThemeModule> {
+		return {
 			ngModule: ThemeModule,
 			providers: [
 				...NbThemeModule.forRoot(
