@@ -1,4 +1,4 @@
-import { Component, OnDestroy, ViewChild } from '@angular/core';
+import { Component, OnDestroy, ViewChild, OnInit } from '@angular/core';
 import { TranslationBaseComponent } from '../../../@shared/language-base/translation-base.component';
 import { Subject } from 'rxjs';
 import { FullCalendarComponent } from '@fullcalendar/angular';
@@ -12,7 +12,9 @@ import { NbDialogService, NbToastrService } from '@nebular/theme';
 import { CandidateInterviewInfoComponent } from '../../../@shared/candidate/candidate-interview-info/candidate-interview-info.component';
 import { CandidateInterviewService } from '../../../@core/services/candidate-interview.service';
 import { CandidateInterviewMutationComponent } from '../../../@shared/candidate/candidate-interview-mutation/candidate-interview-mutation.component';
-import { first } from 'rxjs/operators';
+import { first, takeUntil } from 'rxjs/operators';
+import { CandidatesService } from '../../../@core/services/candidates.service';
+import { Candidate } from '@gauzy/models';
 
 @Component({
 	selector: 'ngx-manage-candidate-interviews',
@@ -20,26 +22,38 @@ import { first } from 'rxjs/operators';
 	styleUrls: ['./manage-candidate-interviews.component.scss']
 })
 export class ManageCandidateInterviewsComponent extends TranslationBaseComponent
-	implements OnDestroy {
+	implements OnInit, OnDestroy {
 	private _ngDestroy$ = new Subject<void>();
 
 	@ViewChild('calendar', { static: true })
 	calendarComponent: FullCalendarComponent;
 	calendarOptions: OptionsInput;
 	selectedInterview = true;
+	selectedCandidateIds: string[];
+	candidates: Candidate[] = [];
 	calendarEvents: EventInput[] = [];
 	constructor(
 		readonly translateService: TranslateService,
 		private dialogService: NbDialogService,
 		private candidateInterviewService: CandidateInterviewService,
-		private toastrService: NbToastrService
+		private toastrService: NbToastrService,
+		private candidatesService: CandidatesService
 	) {
 		super(translateService);
+	}
+	async ngOnInit() {
+		this.candidatesService
+			.getAll(['user'])
+			.pipe(takeUntil(this._ngDestroy$))
+			.subscribe((candidates) => {
+				this.candidates = candidates.items;
+			});
 		this.loadInterviews();
+	}
+	async loadInterviews() {
 		this.calendarOptions = {
 			eventClick: (event) => {
 				const id = event.event._def.extendedProps.id;
-				console.log(event);
 				this.dialogService.open(CandidateInterviewInfoComponent, {
 					context: {
 						interviewId: id
@@ -62,13 +76,11 @@ export class ManageCandidateInterviewsComponent extends TranslationBaseComponent
 			weekends: true,
 			height: 'auto'
 		};
-	}
-
-	async loadInterviews() {
 		const res = await this.candidateInterviewService.getAll([
 			'interviewers'
 		]);
 		if (res) {
+			this.calendarEvents = [];
 			for (const interview of res.items) {
 				this.calendarEvents.push({
 					title: interview.title,
@@ -82,6 +94,14 @@ export class ManageCandidateInterviewsComponent extends TranslationBaseComponent
 			}
 			this.calendarOptions.events = this.calendarEvents;
 		}
+	}
+
+	async onCandidateSelected(id: string) {
+		// TO DO
+		// const candidate = await this.candidatesService.getCandidateById(id, [
+		// 	'user'
+		// ]);
+		// this.selectedCandidate = candidate;
 	}
 	async add() {
 		const dialog = this.dialogService.open(
