@@ -26,6 +26,7 @@ import {
 import { CommandBus } from '@nestjs/cqrs';
 import { CandidateCreateCommand, CandidateBulkCreateCommand } from './commands';
 import { I18nLang } from 'nestjs-i18n';
+import { ParseJsonPipe } from '../shared';
 
 @ApiTags('Candidate')
 @UseGuards(AuthGuard('jwt'))
@@ -89,7 +90,7 @@ export class CandidateController extends CrudController<Candidate> {
 		return this.candidateService.findAll({ where: findInput, relations });
 	}
 
-	@ApiOperation({ summary: 'Find Candidate by id in the same tenant.' })
+	@ApiOperation({ summary: 'Find Candidate by id ' })
 	@ApiResponse({
 		status: HttpStatus.OK,
 		description: 'Found one record',
@@ -100,8 +101,14 @@ export class CandidateController extends CrudController<Candidate> {
 		description: 'Record not found'
 	})
 	@Get(':id')
-	async findById(@Param('id') id: string): Promise<Candidate> {
-		return this.candidateService.findOne(id);
+	async findById(
+		@Param('id') id: string,
+		@Query('data', ParseJsonPipe) data?: any
+	): Promise<Candidate> {
+		const { relations = [] } = data;
+		return this.candidateService.findOne(id, {
+			relations
+		});
 	}
 
 	@ApiOperation({ summary: 'Create new record' })
