@@ -4,22 +4,24 @@ import { ICandidateInterview, Candidate } from '@gauzy/models';
 import { CandidateInterviewersService } from '../../../@core/services/candidate-interviewers.service';
 import { EmployeesService } from '../../../@core/services';
 import { CandidateInterviewService } from '../../../@core/services/candidate-interview.service';
+import { CandidatesService } from '../../../@core/services/candidates.service';
+import { TranslationBaseComponent } from '../../language-base/translation-base.component';
+import { TranslateService } from '@ngx-translate/core';
 @Component({
 	selector: 'ngx-candidate-interview-info',
 	templateUrl: './candidate-interview-info.component.html',
 	styleUrls: ['./candidate-interview-info.component.scss']
 })
-export class CandidateInterviewInfoComponent implements OnInit {
+export class CandidateInterviewInfoComponent extends TranslationBaseComponent
+	implements OnInit {
 	@Input() interviewId: any; //from calendar
 	@Input() interviewList: ICandidateInterview[]; //from profile
-	@Input() selectedCandidate: Candidate;
+	@Input() selectedCandidate: Candidate; //from profile
 	candidateId: string;
 	interviewerNames = [];
 	currentInterview: ICandidateInterview;
-	isMinutes = false;
 	nameList: string;
-	hoursUpdate: number;
-	currentTime: Date = new Date();
+	timeUpdate: string;
 	isNextBtn = true;
 	index = 1;
 	isPreviousBtn = false;
@@ -27,39 +29,35 @@ export class CandidateInterviewInfoComponent implements OnInit {
 		protected dialogRef: NbDialogRef<CandidateInterviewInfoComponent>,
 		private candidateInterviewersService: CandidateInterviewersService,
 		private employeesService: EmployeesService,
+		private candidatesService: CandidatesService,
+		readonly translateService: TranslateService,
 		private candidateInterviewService: CandidateInterviewService
-	) {}
+	) {
+		super(translateService);
+	}
 	closeDialog() {
 		this.dialogRef.close();
 	}
 
-	ngOnInit() {
-		//if (this.interviewId) {
-		//		this.loadInterviews(this.interviewId);
-		//	}
-		this.currentInterview = this.interviewList[0];
-		this.getData(this.currentInterview.id);
-		this.setTime(this.currentInterview.updatedAt);
-	}
-	/*	async loadInterviews(id: string) {
-		if (id) {
-			const res = await this.candidateInterviewService.findById(id);
+	async ngOnInit() {
+		if (this.interviewId) {
+			const res = await this.candidateInterviewService.findById(
+				this.interviewId
+			);
 			if (res) {
 				this.interviewList = [];
 				this.interviewList.push(res);
+
+				const candidate = await this.candidatesService.getCandidateById(
+					res.candidateId,
+					['user']
+				);
+				this.selectedCandidate = candidate;
 			}
 		}
-	}*/
-	setTime(time: any) {
-		const date = new Date(time);
-		this.hoursUpdate = this.currentTime.getHours() - date.getHours();
-		if (this.hoursUpdate === 0) {
-			this.isMinutes = true;
-			this.hoursUpdate =
-				this.currentTime.getMinutes() - date.getMinutes();
-		} else {
-			this.isMinutes = false;
-		}
+		this.currentInterview = this.interviewList[0];
+		this.getData(this.currentInterview.id);
+		this.setTime(this.currentInterview.updatedAt);
 	}
 
 	async getData(id: string) {
@@ -107,6 +105,34 @@ export class CandidateInterviewInfoComponent implements OnInit {
 		this.setTime(this.currentInterview.updatedAt);
 		if (currentIndex === this.interviewList.length - 2) {
 			this.isNextBtn = false;
+		}
+	}
+
+	setTime(time: any) {
+		const now = new Date().getTime();
+		const delta = (now - new Date(time).getTime()) / 1000;
+		if (delta < 60) {
+			this.timeUpdate = this.getTranslation(
+				'CANDIDATES_PAGE.INTERVIEW_INFO_MODAL.LESS_MINUTE'
+			);
+		} else if (delta < 3600) {
+			this.timeUpdate =
+				Math.floor(delta / 60) +
+				this.getTranslation(
+					'CANDIDATES_PAGE.INTERVIEW_INFO_MODAL.MINUTES_AGO'
+				);
+		} else if (delta < 86400) {
+			this.timeUpdate =
+				Math.floor(delta / 3600) +
+				this.getTranslation(
+					'CANDIDATES_PAGE.INTERVIEW_INFO_MODAL.HOURS_AGO'
+				);
+		} else {
+			this.timeUpdate =
+				Math.floor(delta / 86400) +
+				this.getTranslation(
+					'CANDIDATES_PAGE.INTERVIEW_INFO_MODAL.DAYS_AGO'
+				);
 		}
 	}
 }
