@@ -68,6 +68,7 @@ export class SidebarComponent {
 			]
 		}
 	];
+
 	options: ITreeOptions = {
 		isExpandedField: 'expanded',
 		actionMapping: {
@@ -96,7 +97,8 @@ export class SidebarComponent {
 		animateExpand: true,
 		displayField: 'name'
 	};
-	@ViewChild(TreeComponent, { static: false })
+
+	@ViewChild(TreeComponent)
 	private tree: TreeComponent;
 
 	addNode() {
