@@ -87,7 +87,7 @@ const routes: Routes = [
 	},
 	{
 		path: 'server-down',
-		loadChildren: './server-down/server-down.module#ServerDownModule'
+		loadChildren: () => import('./server-down/server-down.module').then(m => m.ServerDownModule)
 	},
 	{ path: 'sign-in/success', component: SignInSuccessComponent },
 	{ path: '', redirectTo: 'pages', pathMatch: 'full' },
