@@ -65,6 +65,7 @@ import { ManageCandidateInterviewsComponent } from './manage-candidate-interview
 import { FullCalendarModule } from '@fullcalendar/angular';
 import { CandidateInterviewInfoModule } from '../../@shared/candidate/candidate-interview-info/candidate-interview-info.module';
 import { CandidateInterviewersService } from '../../@core/services/candidate-interviewers.service';
+import { CandidateMultiSelectModule } from '../../@shared/candidate/candidate-multi-select/candidate-multi-select.module';
 
 export function HttpLoaderFactory(http: HttpClient) {
 	return new TranslateHttpLoader(http, './assets/i18n/', '.json');
@@ -137,7 +138,8 @@ const COMPONENTS = [
 		EmployeeLocationModule,
 		EmployeeRatesModule,
 		StarRatingInputModule,
-		StarRatingOutputModule
+		StarRatingOutputModule,
+		CandidateMultiSelectModule
 	],
 	declarations: [...COMPONENTS],
 	entryComponents: [
