@@ -23,7 +23,7 @@ import { TaskTeamsComponent } from 'apps/gauzy/src/app/@shared/table-components/
 })
 export class TaskComponent extends TranslationBaseComponent
 	implements OnInit, OnDestroy {
-	@ViewChild('tasksTable', { static: false }) tasksTable;
+	@ViewChild('tasksTable') tasksTable;
 	private _ngDestroy$: Subject<void> = new Subject();
 	settingsSmartTable: object;
 	loading = false;
