@@ -8,37 +8,37 @@ import { Environment } from './model';
 import { CloudinaryConfiguration } from '@cloudinary/angular-5.x';
 
 export const environment: Environment = {
-	production: false,
+  production: false,
 
-	API_BASE_URL: 'http://localhost:3000',
-	COMPANY_NAME: 'Ever Co. LTD',
-	COMPANY_SITE: 'Gauzy',
-	COMPANY_LINK: 'https://ever.co/',
-	COMPANY_SITE_LINK: 'https://gauzy.co',
-	COMPANY_GITHUB_LINK: 'https://github.com/ever-co',
-	COMPANY_FACEBOOK_LINK: 'https://www.facebook.com/gauzyplatform',
-	COMPANY_TWITTER_LINK: 'https://twitter.com/gauzyplatform',
-	COMPANY_LINKEDIN_LINK: 'https://www.linkedin.com/company/ever-co.',
-	CLOUDINARY_CLOUD_NAME: 'dv6ezkfxg',
-	CLOUDINARY_API_KEY: '256868982483961',
-	GOOGLE_AUTH_LINK: 'http://localhost:3000/api/auth/google',
-	FACEBOOK_AUTH_LINK: 'http://localhost:3000/api/auth/facebook',
-	LINKEDIN_AUTH_LINK: '#',
-	NO_INTERNET_LOGO: 'assets/images/logos/logo_Gauzy.svg',
-	SENTRY_DNS: 'https://19293d39eaa14d03aac4d3c156c4d30e@sentry.io/4397292',
-	HUBSTAFF_REDIRECT_URI: 'http://localhost:4200/pages/integrations/hubstaff'
+  API_BASE_URL: 'http://localhost:3000',
+  COMPANY_NAME: 'Ever Co. LTD',
+  COMPANY_SITE: 'Gauzy',
+  COMPANY_LINK: 'https://ever.co/',
+  COMPANY_SITE_LINK: 'https://gauzy.co',
+  COMPANY_GITHUB_LINK: 'https://github.com/ever-co',
+  COMPANY_FACEBOOK_LINK: 'https://www.facebook.com/gauzyplatform',
+  COMPANY_TWITTER_LINK: 'https://twitter.com/gauzyplatform',
+  COMPANY_LINKEDIN_LINK: 'https://www.linkedin.com/company/ever-co.',
+  CLOUDINARY_CLOUD_NAME: 'dv6ezkfxg',
+  CLOUDINARY_API_KEY: '256868982483961',
+  GOOGLE_AUTH_LINK: 'http://localhost:3000/api/auth/google',
+  FACEBOOK_AUTH_LINK: 'http://localhost:3000/api/auth/facebook',
+  LINKEDIN_AUTH_LINK: '#',
+  NO_INTERNET_LOGO: 'assets/images/logos/logo_Gauzy.svg',
+  SENTRY_DNS: 'https://19293d39eaa14d03aac4d3c156c4d30e@sentry.io/4397292',
+  HUBSTAFF_REDIRECT_URI: 'http://localhost:4200/pages/integrations/hubstaff'
 };
 
 export const cloudinaryConfiguration: CloudinaryConfiguration = {
-	cloud_name: environment.CLOUDINARY_CLOUD_NAME,
-	api_key: environment.CLOUDINARY_API_KEY
+  cloud_name: environment.CLOUDINARY_CLOUD_NAME,
+  api_key: environment.CLOUDINARY_API_KEY
 };
 
 /*
- * For easier debugging in development mode, you can import the following file
- * to ignore zone related error stack frames such as 'zone.run', 'zoneDelegate.invokeTask'.
- *
- * This import should be commented out in production mode because it will have a negative impact
- * on performance if an error is thrown.
- */
+* For easier debugging in development mode, you can import the following file
+* to ignore zone related error stack frames such as 'zone.run', 'zoneDelegate.invokeTask'.
+*
+* This import should be commented out in production mode because it will have a negative impact
+* on performance if an error is thrown.
+*/
 // import 'zone.js/dist/zone-error';  // Included with Angular CLI.
